# app/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.CompileRequest:
    properties:
      code:
        type: string
      language:
        type: string
    required:
    - code
    - language
    type: object
  dtos.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dtos.ResultResponse:
    properties:
      result:
        type: string
    type: object
...
  dtos.ValidationErrorResponse:
    properties:
      error:
        additionalProperties:
          type: string
        type: object
    type: object
...
host: https://educgpt.ru
info:
  contact: {}
  title: Educ-GPT API
  version: "1.0"
paths:
...
  /utils/compile:
    post:
      consumes:
      - application/json
      description: Compile code by AI
      parameters:
      - description: Bearer <JWT token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Code for compiler
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CompileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Compiled code
          schema:
            $ref: '#/definitions/dtos.ResultResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dtos.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "409":
          description: AI request error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Compile
      tags:
      - utils
swagger: "2.0"